backend/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langchain-community
langchain-text-splitters
langchain-core
faiss-cpu
python-dotenv
tiktoken
pydantic